# get rid of external dependency