--- a/Cargo.toml
+++ b/Cargo.toml
@@ -14,6 +14,3 @@ Froggy is a prototype for the Component Graph System programming model. It aims 
 
 [workspace]
 members = ["demos/cubes"]
-
-[dependencies]
-bit_field = "0.7.0"