This PR gets rid of the `bit_field` dependency.

Benchmark on my local machine:

## Before:
```
test bench_build  ... bench:   1,278,206 ns/iter (+/- 237,475)
test bench_update ... bench:      11,448 ns/iter (+/- 1,920)
...
test bench_build  ... bench:     727,021 ns/iter (+/- 127,100)
test bench_update ... bench:       2,264 ns/iter (+/- 201)
```

## With `bit_field`:

```
test bench_build  ... bench:   1,404,364 ns/iter (+/- 223,314)
test bench_update ... bench:      13,494 ns/iter (+/- 4,488)
...
test bench_build  ... bench:     937,293 ns/iter (+/- 166,870)
test bench_update ... bench:       3,783 ns/iter (+/- 451)
```

## After

```
test bench_build  ... bench:   1,301,554 ns/iter (+/- 422,869)
test bench_update ... bench:       9,692 ns/iter (+/- 1,618)
...
test bench_build  ... bench:     721,164 ns/iter (+/- 109,739)
test bench_update ... bench:       2,177 ns/iter (+/- 458)
```